requirements.txt
# Database
psycopg[binary]==3.2.1
SQLAlchemy>=2.0.0

# Data processing
pandas==2.2.2

# Environment variables
python-dotenv==1.0.1

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Type checking
mypy>=1.5.0
types-psycopg2>=2.9.0

# Code quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Logging
structlog>=23.1.0

# Additional dependencies
tqdm>=4.66.0
typing-extensions>=4.8.0

# Web Framework
flask>=2.3.0

# Vector embeddings
sentence-transformers>=2.2.2
numpy>=1.23.0

# DSPy and HuggingFace integration
dspy-ai>=2.0.0
huggingface_hub>=0.23.0
transformers>=4.38.0
torch>=2.2.0

# Claude API integration
anthropic>=0.21.0

# YAML Parsing for .mdc files
PyYAML \nPyYAML\n
PyYAML
